lan_discovery_WOL: fix udp set_nonblocking twice and lan refactor
